Change viewBox back to width/height parameters to fix crash.
Fix segmentation fault seen on Fritzing 0.7.4b Linux release when
viewBox is used as recommended instead of width/height. Need to
verify if viewBox really should be used to make the connectors
functional and if there's just some parameter issue that causes
the segfault. In the meantime, this allows the svg to at least
be parsed and moved from the part bin.

Signed-off-by: Matt Porter <mporter@ti.com>